Arrumando spin e valor

+++ src/main/java/com/mycompany/supplystoreservice/menu/Produto.java
@@ -26,7 +26,7 @@ public class Produto {
         this.valor = valor;
     }
      
-    public float getId() {
+    public int getId() {
         return id;
     }
 
@@ -42,7 +42,7 @@ public class Produto {
         this.nome = nome;
     }
     
-    public float getQuantidade() {
+    public int getQuantidade() {
         return quantidade;
     }
